refactor: remove live interfaces
